+++ src/components/calcite-checkbox/calcite-checkbox.stories.js
@@ -1,30 +1,38 @@
-import { storiesOf } from '@storybook/html';
-import { withKnobs, boolean, select } from '@storybook/addon-knobs'
-import { darkBackground, parseReadme } from '../../../.storybook/helpers';
-import readme from './readme.md';
+import { storiesOf } from "@storybook/html";
+import { withKnobs, boolean, select } from "@storybook/addon-knobs";
+import { darkBackground, parseReadme } from "../../../.storybook/helpers";
+import readme from "./readme.md";
 const notes = parseReadme(readme);
 
-storiesOf('Checkbox', module)
+storiesOf("Checkbox", module)
   .addDecorator(withKnobs)
-  .add('Simple', () => `
+  .add(
+    "Simple",
+    () => `
     <label>
       <calcite-checkbox
-        checked=${boolean('checked', true)}
-        disabled=${boolean('disabled', false)}
-        indeterminate=${boolean('indeterminate', false)}
-        size=${select('size', {large: 'large', small: 'small'}, 'small')}
+        checked=${boolean("checked", true)}
+        disabled=${boolean("disabled", false)}
+        indeterminate=${boolean("indeterminate", false)}
+        scale="${select("scale", ["s", "m", "l"], "m")}"
       ></calcite-checkbox>
       Text for the checkbox
     </label>
-  `, { notes })
-  .add('Dark mode', () => `
+  `,
+    { notes }
+  )
+  .add(
+    "Dark mode",
+    () => `
     <label>
       <calcite-checkbox
         theme="dark"
-        checked=${boolean('checked', true)}
-        disabled=${boolean('disabled', false)}
-        indeterminate=${boolean('indeterminate', false)}
-        size=${select('size', {large: 'large', small: 'small'}, 'small')}
+        checked=${boolean("checked", true)}
+        disabled=${boolean("disabled", false)}
+        indeterminate=${boolean("indeterminate", false)}
+        scale="${select("scale", ["s", "m", "l"], "m")}"
       ></calcite-checkbox>
     </label>
-`, { notes, backgrounds: darkBackground });
+`,
+    { notes, backgrounds: darkBackground }
+  );