+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Breaking Changes
+
+- `calcite-checkbox` - `size` prop is now `scale` to be consistent with other components
+
 ## [v1.0.0-beta.25] - Apr 28th 2020
 
 ### Breaking Changes